requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2

# AI APIs (optional - for cloud services)
openai==1.3.5
google-generativeai==0.3.0
